% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkbcind-package.R
\docType{package}
\name{nkbcind-package}
\alias{nkbcind}
\alias{nkbcind-package}
\title{nkbcind: Verktyg för kvalitetsindikatorer och populationskarakteristik från NKBC}
\description{
Verktyg för kvalitetsindikatorer och populationskarakteristik från
    Nationellt kvalitetsregister för bröstcancer (NKBC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cancercentrum.bitbucket.io/nkbcind/}
  \item \url{http://bitbucket.org/cancercentrum/nkbcind}
  \item Report bugs at \url{https://bitbucket.org/cancercentrum/nkbcind/issues}
}

}
\author{
\strong{Maintainer}: John Lövrot \email{john.lovrot@sll.se} (\href{https://orcid.org/0000-0002-9339-8059}{ORCID})

Authors:
\itemize{
  \item Lina Benson (\href{https://orcid.org/0000-0002-5393-0055}{ORCID})
}

Other contributors:
\itemize{
  \item RCC [copyright holder, funder]
}

}
\keyword{internal}
